refactor: Quitar mayus de las carpetas de componentes
