+++ src/components/main/Main.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "./Main.css";
-import Present from "../sPresent/Present";
-import Portfolios from "../sPortfolios/Portfolios";
+import Present from "../present/Present";
+import Portfolios from "../portfolios/Portfolios";
 
 function Main() {
   return (
